Use react-query to retrieve knowledge layer user
