+++ package.json
@@ -39,6 +39,7 @@
     "react-hook-form": "^7.37.0",
     "siwe": "^2.1.3",
     "tailwind-merge": "^1.8.0",
+    "tiny-invariant": "^1.3.1",
     "wagmi": "^0.12.10",
     "zod": "^3.21.4"
   },
