+++ src/components/course/create-course-form.tsx
@@ -7,10 +7,10 @@ import { Button } from "@components/basic/button";
 import { Input } from "@components/basic/input";
 import { TextArea } from "@components/basic/textarea/textarea";
 import { FileDropzone } from "@components/file-dropzone";
-import { useKnowledgeLayerContext } from "@context/knowledgelayer-provider";
 import { useCreateProfileModal } from "@hooks/use-create-profile-modal";
 import { useUploadVideos } from "@hooks/use-upload-videos";
 import { useCreateCourse } from "@lib/courses/use-create-course";
+import { useKnowledgeLayerActiveUser } from "@lib/users/use-active-knowledge-layer-user";
 
 import type { Asset } from "@livepeer/react";
 
@@ -27,7 +27,7 @@ interface CreateCourseFields {
 }
 
 export const CreateCourseForm = () => {
-  const { user } = useKnowledgeLayerContext();
+  const { user } = useKnowledgeLayerActiveUser();
   const router = useRouter();
   const openCreateProfileModal = useCreateProfileModal();
