+++ src/lib/courses/use-release-payment.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "wagmi";
 
-import { useKnowledgeLayerContext } from "@context/knowledgelayer-provider";
 import { useKnowledgeLayerEscrow } from "@hooks/use-knowledgelayer-escrow";
+import { useKnowledgeLayerActiveUser } from "@lib/users/use-active-knowledge-layer-user";
 
 import type { ContractReceipt } from "ethers";
 
@@ -14,7 +14,7 @@ interface UseReleasePaymentOptions {
 }
 
 export const useReleasePayment = (options?: UseReleasePaymentOptions) => {
-  const { user } = useKnowledgeLayerContext();
+  const { user } = useKnowledgeLayerActiveUser();
 
   const knowledgeLayerEscrow = useKnowledgeLayerEscrow(true);
   const mutation = useMutation(
